refactor(backoffice): extract useListReorder for template-config dnd
Share the dnd-kit sensor setup and drag-end reorder logic behind a
useListReorder hook. Adopt it in the two configs whose handlers were
the plain reorder shape (image gallery, FAQ items). The section-based
configs reindex order after the move and keep their own handlers.
